+++ ewm-service/src/main/java/ru/practicum/explorewithme/user/service/admin/AdminCategoryService.java
@@ -4,9 +4,9 @@ import ru.practicum.explorewithme.category.model.CategoryRequest;
 import ru.practicum.explorewithme.category.model.CategoryResponse;
 
 public interface AdminCategoryService {
-    void deleteCategory(Long catId);
+    void deleteCategory(Integer catId);
 
     CategoryResponse createCategory(CategoryRequest category);
 
-    CategoryResponse updateCategory(CategoryRequest category, Long catId);
+    CategoryResponse updateCategory(CategoryRequest category, Integer catId);
 }
